refactor: migrate Comission model to organization app
Migrated the model to organization seeing as there is no reason
to have a whole django app for a single model and almost no logic.
The app will be deprecated in the future. The migration requried
the following steps

*copy over model source code to organization.models
*update references in factories, app admins and models import Comission
*generating the migration
