+++ commissions/admin.py
@@ -1,5 +1,2 @@
-from django.contrib import admin
 
-from commissions.models import Commission
 
-admin.site.register(Commission)
