+++ organization/models.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 from django.db import models
 from users.models import User
 
+
 class InternalGroup(models.Model):
     class Meta:
         default_related_name = 'groups'
@@ -56,3 +57,18 @@ class InternalGroupPosition(models.Model):
 
     def __repr__(self):
         return "Position(name=%s)" % self.name
+
+
+class Commission(models.Model):
+    """
+    A commissions (verv) in KSG. A commissions can be shared by many users (e.g. Personal),
+    or created specifically for this internal group (e.g. Hybelbarsjef)
+    """
+
+    name = models.CharField(max_length=32, unique=True)
+
+    def __str__(self):
+        return "Commission %s" % (self.name,)
+
+    def __repr__(self):
+        return "Commission(name=%s)" % (self.name,)
